+++ packages/resolver.go
@@ -7,7 +7,7 @@ package packages
 import "net/http"
 
 type RemoteResolver interface {
-	RedirectArtifactsHandler(w http.ResponseWriter, r *http.Request, p *Package)
-	RedirectStaticHandler(w http.ResponseWriter, r *http.Request, p *Package, resourcePath string)
-	RedirectSignaturesHandler(w http.ResponseWriter, r *http.Request, p *Package)
+	ArtifactsHandler(w http.ResponseWriter, r *http.Request, p *Package)
+	StaticHandler(w http.ResponseWriter, r *http.Request, p *Package, resourcePath string)
+	SignaturesHandler(w http.ResponseWriter, r *http.Request, p *Package)
 }