+++ storage/fakestorage.go
@@ -17,7 +17,6 @@ const fakePackageStorageBucketInternal = "fake-package-storage-internal"
 
 var FakeIndexerOptions = IndexerOptions{
 	PackageStorageBucketInternal: "gs://" + fakePackageStorageBucketInternal,
-	PackageStorageEndpoint:       "https://package-storage.elastic.co/",
 	WatchInterval:                0,
 }